requirements.txt
Flask==2.2.2
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
python-dotenv==0.21.0